% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{cw_columns_en_de}
\alias{cw_columns_en_de}
\title{Crosswalk column names: \code{en} to \code{de}}
\usage{
cw_columns_en_de(
  data,
  path_crosswalk_def = app_sys("app/config_crosswalk.yml")
)
}
\arguments{
\item{data}{}
}
\description{
Crosswalk column names: \code{en} to \code{de}
}
