% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{cw_encode_column}
\alias{cw_encode_column}
\title{Crosswalk: encode column}
\usage{
cw_encode_column(data, path_cw_encoding, col = character(), vec_cast = TRUE)
}
\arguments{
\item{data}{\code{\link[=tibble]{tibble()}}}

\item{vec_cast}{}
}
\description{
Crosswalk: encode column
}
