% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{cw_encoding_expand}
\alias{cw_encoding_expand}
\title{Crosswalk: expand encoding}
\usage{
cw_encoding_expand(
  encoding,
  language = "en",
  encoding_add = c("de_int", "de_ext"),
  encoding_prefill = c("NA", "original")
)
}
\arguments{
\item{overwrite}{}
}
\description{
Crosswalk: expand encoding
}
