% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{cw_encoding_extract}
\alias{cw_encoding_extract}
\title{Crosswalk: extract encoding}
\usage{
cw_encoding_extract(data)
}
\arguments{
\item{data}{\code{\link[=tibble]{tibble()}}}
}
\description{
Crosswalk: extract encoding
}
