% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{cw_encoding_persist}
\alias{cw_encoding_persist}
\title{Crosswalk: persist encoding}
\usage{
cw_encoding_persist(
  encoding,
  path_dir = here::here("datalake", "layer_01"),
  overwrite = FALSE
)
}
\arguments{
\item{overwrite}{}
}
\description{
Crosswalk: persist encoding
}
