% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{cw_encodings_expand}
\alias{cw_encodings_expand}
\title{Crosswalk: expand encodings}
\usage{
cw_encodings_expand(
  encodings,
  language = "en",
  encoding_add = c("de_int", "de_ext")
)
}
\arguments{
\item{encodings}{DESCRIPTION.}

\item{language}{DESCRIPTION.}

\item{encoding_add}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Expand encodings
}
\examples{
# ADD_EXAMPLES_HERE
}
