% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{cw_encodings_persist}
\alias{cw_encodings_persist}
\title{Persist encoding crosswalk files}
\usage{
cw_encodings_persist(
  encodings,
  path_dir = here::here("datalake", "layer_01"),
  overwrite = FALSE
)
}
\arguments{
\item{overwrite}{}
}
\description{
Persist encoding crosswalk files
}
