% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{cw_rename_columns}
\alias{cw_rename_columns}
\title{FUNCTION_TITLE}
\usage{
cw_rename_columns(data, crosswalk_table, from = "internal", to = "external_en")
}
\arguments{
\item{data}{DESCRIPTION.}

\item{crosswalk_table}{DESCRIPTION.}

\item{from}{DESCRIPTION.}

\item{to}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
