% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbt.def.data-package.R
\docType{package}
\name{dbt.def.data-package}
\alias{dbt.def.data}
\alias{dbt.def.data-package}
\title{dbt.def.data: Datasets for DBT DEF Training}
\description{
Datasets for DBT DEF Training
}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io}

}
\keyword{internal}
