% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pkg_file}
\alias{pkg_file}
\title{Path to package file}
\usage{
pkg_file(path)
}
\arguments{
\item{path}{\link{character} Path to directory or file}
}
\value{
\link{character} Full path to directory of file
}
\description{
Path to package file.
}
\examples{
pkg_file("DESCRIPTION")
}
